feature: view function in evaluation loop
